fix intersection filter flow
When using the intersection filter, if one of the search returned no result, it would incorrectly ignore it.

For example, if you searched with song name and artist name intersection, but the song name did not exist at all in the database, it would ignore it entirely and you would get every song by the artist instead of an empty result.